requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
feedgen==0.9.0
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
lxml==4.9.2
python-dateutil==2.8.2
requests==2.28.2
six==1.16.0
smmap==5.0.0
urllib3==1.26.15
